alioth: sepolicy: allow parts to read default_prop
03-28 18:41:35.003  3294  3294 W eageos.settings: type=1400 audit(0.0:243): avc:  denied  { read } for  name="u:object_r:default_prop:s0" dev="tmpfs" ino=2192 scontext=u:r:devicesettings_app:s0 tcontext=u:object_r:default_prop:s0 tclass=file permissive=0

alioth: sepolicy: allow parts to get SettingsLib prop

ThermalSettingsFragment's use of ApplicationsState includes a
check of whether a package is an instant app or not. The function
isInstant() in AppUtils reads settingsdebug.instant.packages prop
to retrieve a list of instant packages for debugging purposes.

We do not actually use this property, but this check triggers an
SELinux denial for every package in the list and spams the logs
with permission denied errors. Get rid of these log spams by
allowing parts to read the prop.

Test: open the thermal profile settings and check logs

alioth: fix! sepolicy for hbm
Co-authored-by: Sanjivns <113604406+Sanjivns@users.noreply.github.com>